Move Dashboard-specific modals to Dashboard/modals
