+++ src/components/Dashboard/Pay/index.tsx
@@ -1,8 +1,8 @@
 import { Button, Tooltip } from 'antd'
 import { t, Trans } from '@lingui/macro'
 
-import ConfirmPayOwnerModal from 'components/modals/ConfirmPayOwnerModal'
-import PayWarningModal from 'components/modals/PayWarningModal'
+import ConfirmPayOwnerModal from 'components/Dashboard/modals/ConfirmPayOwnerModal'
+import PayWarningModal from 'components/Dashboard/modals/PayWarningModal'
 import InputAccessoryButton from 'components/shared/InputAccessoryButton'
 import FormattedNumberInput from 'components/shared/inputs/FormattedNumberInput'
 import { V1ProjectContext } from 'contexts/v1/projectContext'
