+++ src/components/Dashboard/Rewards.tsx
@@ -1,9 +1,9 @@
 import { BigNumber } from '@ethersproject/bignumber'
 import { t, Trans } from '@lingui/macro'
 import { Button, Descriptions, Modal, Space, Statistic, Tooltip } from 'antd'
-import ConfirmUnstakeTokensModal from 'components/modals/ConfirmUnstakeTokensModal'
-import ParticipantsModal from 'components/modals/ParticipantsModal'
-import RedeemModal from 'components/modals/RedeemModal'
+import ConfirmUnstakeTokensModal from 'components/Dashboard/modals/ConfirmUnstakeTokensModal'
+import ParticipantsModal from 'components/Dashboard/modals/ParticipantsModal'
+import RedeemModal from 'components/Dashboard/modals/RedeemModal'
 import FormattedAddress from 'components/shared/FormattedAddress'
 import { NetworkContext } from 'contexts/networkContext'
 import { V1ProjectContext } from 'contexts/v1/projectContext'
@@ -23,7 +23,7 @@ import React, { CSSProperties, useContext, useState } from 'react'
 import { formatPercent, formatWad } from 'utils/formatNumber'
 import { decodeFundingCycleMetadata } from 'utils/fundingCycle'
 
-import PrintPreminedModal from '../modals/PrintPreminedModal'
+import PrintPreminedModal from './modals/PrintPreminedModal'
 import IssueTickets from './IssueTickets'
 import SectionHeader from './SectionHeader'
