+++ src/components/Dashboard/Spending.tsx
@@ -2,7 +2,7 @@ import { CrownFilled } from '@ant-design/icons'
 import { t, Trans } from '@lingui/macro'
 
 import { Button, Space, Tooltip } from 'antd'
-import WithdrawModal from 'components/modals/WithdrawModal'
+import WithdrawModal from 'components/Dashboard/modals/WithdrawModal'
 import Balance from 'components/Navbar/Balance'
 import CurrencySymbol from 'components/shared/CurrencySymbol'
 import TooltipLabel from 'components/shared/TooltipLabel'
