+++ src/components/shared/RichNote.tsx
